refactor(sdk): Move Dialoguer AuthHelper to auth-core
dialoguer forced every sdk_core consumer (including headless/server
users) to compile a TTY prompt stack. Move Dialoguer impl into
openstack-sdk-auth-core (already a dependency, no cycle), gated
behind a new `interactive` feature (default-on, forwarded from
sdk_core). sdk_core's auth_helper module re-exports it, so all
downstream import paths stay unchanged.

Also drop tokio-util (zero references in sdk_core) and inventory
(dead there; real usage lives in auth-core/auth-* plugin crates),
and move tokio to dev-dependencies (only used by #[tokio::test]).

Removed OpenStackError::Dialoguer variant: never constructed anywhere
in the workspace.
